refactor: improve Supabase directory structure and add syncSupabase router function
