+++ src/api/supabase/fetch/index.ts
@@ -1,0 +1,13 @@
+import { fetchArticlesFromSupabase } from './fetchArticles';
+import { fetchCommentsFromSupabase } from './fetchComments';
+import { fetchDetailsFromSupabase } from './fetchDetails';
+import { fetchProfilesFromSupabase } from './fetchProfiles';
+import { fetchTagListFromSupabase } from './fetchTags';
+
+export {
+    fetchArticlesFromSupabase,
+    fetchCommentsFromSupabase,
+    fetchDetailsFromSupabase,
+    fetchProfilesFromSupabase,
+    fetchTagListFromSupabase,
+};
